backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
asyncpg==0.29.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
httpx==0.27.2
python-multipart==0.0.9
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-httpx==0.30.0
